+++ Sources/Behaviours/KeyboardStickyBehaviour.swift
@@ -1,9 +1,8 @@
 //
 // AppwiseCore
-// Copyright © 2022 Appwise
+// Copyright © 2023 Wisemen
 //
 
-import AppwiseCore
 import UIKit
 
 /// Tries to ensure that a scrollview grows to fill the screen, but if a