+++ Sources/CoreData/Repository/Repository.swift
@@ -1,8 +1,11 @@
 //
 // AppwiseCore
-// Copyright © 2022 Appwise
+// Copyright © 2023 Wisemen
 //
 
 /// Put all your repositories in this namespace
 public enum Repository {
 }
+
+/// Internal error thrown by (for example) the `refresh()` functions.
+struct UnimplementedMethod: Error {}