+++ Sources/UI/TintedImageView.swift
@@ -1,6 +1,6 @@
 //
 // AppwiseCore
-// Copyright © 2022 Appwise
+// Copyright © 2023 Wisemen
 //
 
 import IBAnimatable
@@ -9,7 +9,7 @@ open class TintedImageView: AnimatableImageView {
 	override open func awakeFromNib() {
 		super.awakeFromNib()
 
-		// swiftlint:disable swiftgen_colors
+		// swiftlint:disable:next swiftgen_colors
 		let tint = tintColor
 		tintColor = #colorLiteral(red: 0.1, green: 0.2, blue: 0.3, alpha: 1)
 		tintColor = tint