+++ Sources/UIApplication/Application Services/AddSkipBackupAttributeApplicationService.swift
@@ -1,6 +1,6 @@
 //
 // AppwiseCore
-// Copyright © 2022 Appwise
+// Copyright © 2023 Wisemen
 //
 
 import CocoaLumberjack
@@ -28,6 +28,7 @@ final class AddSkipBackupAttributeApplicationService: NSObject, ApplicationServi
 			}
 
 			// add skip attribute
+			// swiftlint:disable:next legacy_objc_type
 			try (supportDirectory as NSURL).setResourceValue(true, forKey: URLResourceKey.isExcludedFromBackupKey)
 		} catch {
 			DDLogError("Error excluding support directory from backup: \(error)")